+++ src/collective/easytag/testing.py
@@ -1,10 +1,12 @@
+# pylint: disable=W0613
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import FunctionalTesting
+# from plone.app.testing import FunctionalTesting
 from plone.app.testing import applyProfile
 
 from zope.configuration import xmlconfig
+
 
 class CollectiveEasytag(PloneSandboxLayer):
 
@@ -16,7 +18,6 @@ class CollectiveEasytag(PloneSandboxLayer):
         xmlconfig.file('configure.zcml',
                        collective.easytag,
                        context=configurationContext)
-
 
     def setUpPloneSite(self, portal):
         applyProfile(portal, 'collective.easytag:default')